+++ test/smokes/go_vet/success/main.go
@@ -1,8 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-	awesome_text := "Hello World!"
-	fmt.Printf("text", awesome_text)
-}