+++ test/smokes/golint/success/main.go
@@ -1,8 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-	awesome_text := "Hello World!"
-	fmt.Print(awesome_text)
-}