+++ test/smokes/gometalinter/import_path_and_install_path/cat/cat.go
@@ -1,8 +1,0 @@
-package cat
-
-import "fmt"
-
-func Meow(s string) {
-	fmt.Print("Meow!")
-	fmt.Printf(s)
-}