+++ test/smokes/gometalinter/install_path/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/sideci/install_path_test/cat"
-
-func main() {
-	cat.Meow("Nyan")
-}