+++ test/smokes/gometalinter/success/main.go
@@ -1,9 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-	if true == true {
-		fmt.Print("Hello World!")
-	}
-}